✨ Separate fuzz test template generator, update generator logic and update tests
