+++ crates/template/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2021"
 [dependencies]
 
 # Trident IDL spec
-trident-idl-spec = "0.0.1"
+trident-idl-spec = { git = "https://github.com/Ackee-Blockchain/trident-idl-spec", version = "0.0.1" }
 
 # Misc
 syn = { version = "2", features = ["visit", "full"] }
